create dedicated test suite for insights by period repository
