+++ packages/cli/src/modules/insights/database/repositories/__tests__/insights-by-period.repository.test.ts
@@ -1,0 +1,77 @@
+import { Container } from '@n8n/di';
+import { DateTime } from 'luxon';
+
+import { InsightsConfig } from '@/modules/insights/insights.config';
+import { createTeamProject } from '@test-integration/db/projects';
+import { createWorkflow } from '@test-integration/db/workflows';
+import * as testDb from '@test-integration/test-db';
+
+import { createCompactedInsightsEvent, createMetadata } from '../../entities/__tests__/db-utils';
+import { InsightsByPeriodRepository } from '../insights-by-period.repository';
+
+describe('InsightsByPeriodRepository', () => {
+	beforeAll(async () => {
+		await testDb.init();
+	});
+
+	describe('Avoid deadlock error', () => {
+		let defaultBatchSize: number;
+		beforeAll(() => {
+			// Store the original config value
+			const insightsConfig = Container.get(InsightsConfig);
+			defaultBatchSize = insightsConfig.compactionBatchSize;
+
+			// Set a smaller batch size to trigger the deadlock error
+			insightsConfig.compactionBatchSize = 3;
+		});
+
+		afterAll(() => {
+			// Reset the config to its original state
+			const insightsConfig = Container.get(InsightsConfig);
+			insightsConfig.compactionBatchSize = defaultBatchSize;
+		});
+
+		test('should not throw deadlock error on concurrent compaction', async () => {
+			// ARRANGE
+			const insightsConfig = Container.get(InsightsConfig);
+			const insightsByPeriodRepository = Container.get(InsightsByPeriodRepository);
+			const transactionSpy = jest.spyOn(insightsByPeriodRepository.manager, 'transaction');
+			const project = await createTeamProject();
+			const workflow = await createWorkflow({}, project);
+			await createMetadata(workflow);
+
+			const batchQuery = insightsByPeriodRepository.getPeriodInsightsBatchQuery({
+				periodUnitToCompactFrom: 'hour',
+				compactionBatchSize: insightsConfig.compactionBatchSize,
+				maxAgeInDays: insightsConfig.compactionHourlyToDailyThresholdDays,
+			});
+
+			// Create test data
+			const promises = [];
+			for (let i = 0; i < 100; i++) {
+				await createCompactedInsightsEvent(workflow, {
+					type: 'success',
+					value: 1,
+					periodUnit: 'hour',
+					periodStart: DateTime.now().minus({ day: 91, hour: i + 1 }),
+				});
+			}
+
+			// ACT
+			for (let i = 0; i < 10; i++) {
+				promises.push(
+					insightsByPeriodRepository.compactSourceDataIntoInsightPeriod({
+						sourceBatchQuery: batchQuery,
+						sourceTableName: insightsByPeriodRepository.metadata.tableName,
+						periodUnitToCompactInto: 'day',
+					}),
+				);
+			}
+
+			// ASSERT
+			// await all promises concurrently
+			await expect(Promise.all(promises)).resolves.toBeDefined();
+			expect(transactionSpy).toHaveBeenCalledTimes(1);
+		});
+	});
+});
